Port Haddock.Interface to new GHC API.
This required one bigger change: 'readInterfaceFile' used to take an
optional 'Session' argument.  This was used to optionally update the
name cache of an existing GHC session.  This does not work with the
new GHC API, because an active session requires the function to return
a 'GhcMonad' action, but this is not possible if no session is
provided.

The solution is to use an argument of functions for reading and
updating the name cache and to make the function work for any monad
that embeds IO, so it's result type can adapt to the calling context.

While refactoring, I tried to make the code a little more
self-documenting, mostly turning comments into function names.