Fix bad error handling with newer GHCs
When support for GHC 6.10 was added, an error handler was installed only around
the typechecking phase. This had the effect that errors thrown during
dependency chasing were caught in the top-level exception handler and not
printed with enough detail.  With this patch we wrap the error handler around
all our usage of the Ghc monad.